% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_to_ct.R
\name{deparse_query}
\alias{deparse_query}
\title{extract variables and associated do-operations from query
and re-write query in terms of realized outcomes post do-operation}
\usage{
deparse_query(query, join_by, nodes)
}
\arguments{
\item{query}{String specifying query}

\item{join_by}{Logical operator. Used to connect causal statements: \emph{AND} ('&') or \emph{OR} ('|').}

\item{nodes}{Character String of nodes in model}
}
\value{
A List containing, the reconstructed query, a list of do-operations, a vector of node names attached to the dos, a vector of newly constructed variables
}
\description{
extract variables and associated do-operations from query
and re-write query in terms of realized outcomes post do-operation
}
\keyword{internal}
