% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_type_prob_cstd}
\alias{get_type_prob_cstd}
\title{generates one draw from type probability distribution for each type in P}
\usage{
get_type_prob_cstd(P, parameters, nrow, ncol)
}
\arguments{
\item{P}{parameter_matrix of parameters and causal types}

\item{parameters, }{priors or posteriors}

\item{nrow}{number of rows in P}

\item{ncol}{number of columns in P}
}
\value{
draw from type distribution for each type in P
}
\description{
generates one draw from type probability distribution for each type in P
}
