% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_type_prob_multiple_cstd}
\alias{get_type_prob_multiple_cstd}
\title{generates n draws from type probability distribution for each type in P}
\usage{
get_type_prob_multiple_cstd(params, P, nrow, ncol, nrow_p, ncol_p)
}
\arguments{
\item{params}{parameters, priors or posteriors}

\item{P}{parameter_matrix of parameters and causal types}

\item{nrow}{number of rows in params matrix before being flattend to vector}

\item{ncol}{numbers of columns in params matrix before being flattened to vector}

\item{nrow_p}{number of rows in P matrix before being flattened to vector}

\item{ncol_p}{number of columns in P matrix befroe being flattened to vector}
}
\value{
draws from type distribution for each type in P
}
\description{
generates n draws from type probability distribution for each type in P
}
