% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{interact_model}
\alias{interact_model}
\title{helper to add interactions on selected nodes to a \code{causal_model} made with \link{make_simple_model}}
\usage{
interact_model(model, interactions, interaction_operation = c("&", "|"))
}
\arguments{
\item{model}{\code{causal_model} made with \link{make_simple_model}}

\item{interactions}{a list of character vectors specifying nodes to interact (the first 1:n-1 nodes per vector specify the interacting nodes, the nth node specifies the interaction target node)}

\item{interaction_operation}{character vector of logical operations to create interactions with}
}
\value{
An object of class \code{causal_model} with the specified interactions
}
\description{
helper to add interactions on selected nodes to a \code{causal_model} made with \link{make_simple_model}
}
\examples{
\dontrun{

model <- make_simple_model("X -> M1; X -> M2; M1 -> Y; M2 -> Y") |>
            interact_model(interactions = list(c("M1","M2","Y")))

}
}
