% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_causal_types_c}
\alias{make_causal_types_c}
\title{cpp helper to make causal types}
\usage{
make_causal_types_c(nodal_types)
}
\arguments{
\item{nodal_types}{a List of nodal types}
}
\value{
vector of vectors containing causal types
}
\description{
cpp helper to make causal types
}
\keyword{internal}
