% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{make_simple_model}
\alias{make_simple_model}
\title{helper to make model with strictly monotonic relationships and no interactions}
\usage{
make_simple_model(statement)
}
\arguments{
\item{statement}{string describing causal model}
}
\value{
An object of class \code{causal_model}.
}
\description{
helper to make model with strictly monotonic relationships and no interactions
}
\examples{
\dontrun{

model <- make_simple_model("X -> M1; X -> M2; M1 -> Y; M2 -> Y")

}
}
