% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_to_ct.R
\name{query_to_ct}
\alias{query_to_ct}
\title{map query to causal types. Finds causal types associated with query,
writes them to a bigmatrix and returns a pointer to the assocaited
memory address.}
\usage{
query_to_ct(model, query, join_by = "|", file_name = "")
}
\arguments{
\item{model}{a causal model}

\item{query}{string specifying query}

\item{join_by}{Logical operator. Used to connect causal statements: \emph{AND} ('&') or \emph{OR} ('|').}

\item{file_name}{file name for the file backed bigmatrix object}
}
\value{
pointer to a bigmatrix memory location
}
\description{
map query to causal types. Finds causal types associated with query,
writes them to a bigmatrix and returns a pointer to the assocaited
memory address.
}
