% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{query_to_ct_c}
\alias{query_to_ct_c}
\title{cpp implementation of realise_outcomes for map_query_to_causal_types. Dos are evaluated
and the realised outcomes for the variable they are attached to is written to a bigmatrix.}
\usage{
query_to_ct_c(
  nodes,
  endogenous_nodes,
  dos,
  parents_list,
  nodal_types,
  nodal_types_colnames,
  nodal_types_collapsed,
  n_causal_types,
  vars
)
}
\arguments{
\item{nodes}{string vector of nodes names}

\item{endogenous_nodes}{string vector of endogenous nodes}

\item{dos}{List of do operations}

\item{parents_list}{List of parent nodes for each node}

\item{nodal_types}{List of integer matrices with uncollapsed nodal types}

\item{nodal_types_colnames}{List of column names of matrices in nodal_types}

\item{nodal_types_collapsed}{List of collapsed nodal types}

\item{n_causal_types}{int specifying number of causal types}

\item{vars}{string vector with names of variables dos are attached to}
}
\description{
cpp implementation of realise_outcomes for map_query_to_causal_types. Dos are evaluated
and the realised outcomes for the variable they are attached to is written to a bigmatrix.
}
