% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{realise_outcomes_c}
\alias{realise_outcomes_c}
\title{cpp implementation of realise_outcomes. Realise outcomes for all causal types.
Calculated by sequentially calculating endogenous nodes. If a do operator is applied to
any node then it takes the given value and all its descendants are generated accordingly.
Output is written to a bigmatrix.}
\usage{
realise_outcomes_c(
  outcomes,
  nodes,
  endogenous_nodes,
  dos,
  parents_list,
  nodal_types,
  nodal_types_colnames,
  nodal_types_collapsed,
  n_causal_types
)
}
\arguments{
\item{outcomes}{memory address of a bigmatrix object}

\item{nodes}{string vector of nodes names}

\item{endogenous_nodes}{string vector of endogenous nodes}

\item{dos}{List of do operations}

\item{parents_list}{List of parent nodes for each node}

\item{nodal_types}{List of integer matrices with uncollapsed nodal types}

\item{nodal_types_colnames}{List of column names of matrices in nodal_types}

\item{nodal_types_collapsed}{List of collapsed nodal types}

\item{n_causal_types}{int specifying number of causal types}
}
\description{
cpp implementation of realise_outcomes. Realise outcomes for all causal types.
Calculated by sequentially calculating endogenous nodes. If a do operator is applied to
any node then it takes the given value and all its descendants are generated accordingly.
Output is written to a bigmatrix.
}
