% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{simple_nodes}
\alias{simple_nodes}
\title{helper to make monotonic non-interacting nodal types for a node with n parents}
\usage{
simple_nodes(n = 5)
}
\arguments{
\item{n}{integer specifying number of parents}
}
\value{
a character vector of nodal types
}
\description{
helper to make monotonic non-interacting nodal types for a node with n parents
}
\examples{
\dontrun{

simple_nodes(n = 5)

}
}
\keyword{internal}
