% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{split}
\alias{split}
\title{split a DAG into all component submodels by running submodel across each child}
\usage{
split(model, add_causal_types = FALSE, trivial = FALSE)
}
\arguments{
\item{model}{A \code{causal_model}}

\item{add_causal_types}{logical specifying whether to add causal types to the submodel. Defaults to \code{FALSE}.}

\item{trivial}{logical indicating whether to include trivial single node submodels for exogenous nodes. Defaults to \code{FALSE}.}
}
\value{
a list of submodels of type \code{causal_model} for each child node.
}
\description{
split a DAG into all component submodels by running submodel across each child
}
\examples{
\dontrun{

model <- make_simple_model("X -> M1; X -> M2; M1 -> Y; M2 -> Y")
split(model)

}
}
\keyword{internal}
