% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{submodel}
\alias{submodel}
\title{create a submodel of a DAG given a child node}
\usage{
submodel(model, child, add_causal_types = FALSE)
}
\arguments{
\item{model}{A \code{causal_model}}

\item{child}{character string specifying the name of the child node for which to create a submodel}

\item{add_causal_types}{logical specifying whether to add causal types to the submodel. Defaults to \code{FALSE}.}
}
\value{
\code{causal_model} encoding submodel for \code{child}.
}
\description{
create a submodel of a DAG given a child node
}
\examples{
\dontrun{

model <- make_simple_model("X -> M1; X -> M2; M1 -> Y; M2 -> Y")
submodel(model,"M1")

}
}
\keyword{internal}
