% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncollapse_nt.R
\name{uncollapse_nt}
\alias{uncollapse_nt}
\title{helper to uncollapse nodal types}
\usage{
uncollapse_nt(nodal_types)
}
\arguments{
\item{nodal_tpyes}{a List of nodal types}
}
\value{
A List of numeric matrices representing uncollapsed nodal types
}
\description{
helper to uncollapse nodal types
}
\keyword{internal}
