% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{update_stitch}
\alias{update_stitch}
\title{Fit causal model using 'stan' by updating over submodels with nested parallelism and stitching the results back into a fully updated complete causal model.}
\usage{
update_stitch(
  model,
  data,
  chains = 4L,
  parallel = TRUE,
  verbose = FALSE,
  trivial = FALSE
)
}
\arguments{
\item{model}{A \code{causal_model}}

\item{data}{A \code{data.frame}.  Data of nodes that can take three values: 0, 1, and NA. In long form as generated by \code{\link{make_events}}}

\item{chains}{integer specifying the number of markov chains to run for each 'stan' model. Defaults to 4.}

\item{parallel}{logical indicating whether to parallelize across submodels. If \code{FALSE} submodels are updated sequentially with markov chains running in parallel for each submodel. If \code{TRUE} an optimal nested parallel evaluation strategy is generated given available cores and submodels as well as markov chains within submodels are run in parallel. Defaults to \code{TRUE}.}

\item{verbose}{logical indicating whether to print intermediate output from 'stan' to console. Defaults to \code{FALSE}.}

\item{trivial}{logical indicating whether to update trivial single node submodels for exogenous nodes. This causes redundant updating and is not necessary for standard use. Defaults to \code{FALSE}.}
}
\value{
an updated \code{causal_model}
}
\description{
Fit causal model using 'stan' by updating over submodels with nested parallelism and stitching the results back into a fully updated complete causal model.
}
\examples{
\dontrun{

model <- make_simple_model("X -> M1; X -> M2; M1 -> Y; M2 -> Y")
data <- CausalQueries::simulate_data(model,300)

model <- update_stitch(model, data, chains = 1L, parallel = TRUE)

}
}
